+++ NINA.WPF.Base/ViewModel/Equipment/Telescope/TelescopeVM.cs
@@ -403,22 +403,28 @@ namespace NINA.WPF.Base.ViewModel.Equipment.Telescope {
                             if (Math.Abs(Telescope.SiteLatitude - profileService.ActiveProfile.AstrometrySettings.Latitude) > LAT_LONG_TOLERANCE
                                 || Math.Abs(Telescope.SiteLongitude - profileService.ActiveProfile.AstrometrySettings.Longitude) > LAT_LONG_TOLERANCE
                                 || Math.Abs(Telescope.SiteElevation - profileService.ActiveProfile.AstrometrySettings.Elevation) >= SITE_ELEVATION_TOLERANCE) {
-                                var syncVM = new TelescopeLatLongSyncVM(
-                                    profileService.ActiveProfile.AstrometrySettings.Latitude,
-                                    profileService.ActiveProfile.AstrometrySettings.Longitude,
-                                    profileService.ActiveProfile.AstrometrySettings.Elevation,
-                                    Telescope.SiteLatitude,
-                                    Telescope.SiteLongitude,
-                                    Telescope.SiteElevation
-                                );
-                                await WindowService.ShowDialog(syncVM, Loc.Instance["LblSyncLatLong"], System.Windows.ResizeMode.NoResize, System.Windows.WindowStyle.ToolWindow);
 
-                                if (syncVM.Mode == LatLongSyncMode.NINA) {
+                                TelescopeLocationSyncDirection syncMode = profileService.ActiveProfile.TelescopeSettings.TelescopeLocationSyncDirection;
+                                if(profileService.ActiveProfile.TelescopeSettings.TelescopeLocationSyncDirection == TelescopeLocationSyncDirection.PROMPT) { 
+                                    var syncVM = new TelescopeLatLongSyncVM(
+                                        profileService.ActiveProfile.AstrometrySettings.Latitude,
+                                        profileService.ActiveProfile.AstrometrySettings.Longitude,
+                                        profileService.ActiveProfile.AstrometrySettings.Elevation,
+                                        Telescope.SiteLatitude,
+                                        Telescope.SiteLongitude,
+                                        Telescope.SiteElevation
+                                    );
+                                    await WindowService.ShowDialog(syncVM, Loc.Instance["LblSyncLatLong"], System.Windows.ResizeMode.NoResize, System.Windows.WindowStyle.ToolWindow);
+
+                                    syncMode = syncVM.Mode;
+                                } 
+
+                                if (syncMode == TelescopeLocationSyncDirection.TOAPPLICATION) {
                                     Logger.Info($"Importing coordinates from mount into N.I.N.A. - Mount latitude {Telescope.SiteLatitude} , longitude {Telescope.SiteLongitude}, elevation {Telescope.SiteElevation} -> N.I.N.A. latitude {profileService.ActiveProfile.AstrometrySettings.Latitude} , longitude {profileService.ActiveProfile.AstrometrySettings.Longitude}, elevation {profileService.ActiveProfile.AstrometrySettings.Elevation}");
                                     profileService.ChangeLatitude(Telescope.SiteLatitude);
                                     profileService.ChangeLongitude(Telescope.SiteLongitude);
                                     profileService.ChangeElevation(Telescope.SiteElevation);
-                                } else if (syncVM.Mode == LatLongSyncMode.TELESCOPE) {
+                                } else if (syncMode == TelescopeLocationSyncDirection.TOTELESCOPE) {
                                     Logger.Info($"Importing coordinates from N.I.N.A. into Mount - N.I.N.A. latitude {profileService.ActiveProfile.AstrometrySettings.Latitude} , longitude {profileService.ActiveProfile.AstrometrySettings.Longitude}, elevation {profileService.ActiveProfile.AstrometrySettings.Elevation} -> Mount latitude {Telescope.SiteLatitude} , longitude {Telescope.SiteLongitude}, elevation {Telescope.SiteElevation}");
                                     var targetLatitude = profileService.ActiveProfile.AstrometrySettings.Latitude;
                                     var targetLongitude = profileService.ActiveProfile.AstrometrySettings.Longitude;
@@ -433,6 +439,8 @@ namespace NINA.WPF.Base.ViewModel.Equipment.Telescope {
                                         Notification.ShowError(string.Format(Loc.Instance["LblUnableToSetMountLatLong"], Math.Round(targetLatitude, 3), Math.Round(targetLongitude, 3)));
                                     }
                                 }
+                                
+
                             }
 
                             TelescopeInfo.CopyFrom(new TelescopeInfo {